### 2021-04-22 #3206 @whedon
[PRE REVIEW]: JPhyloRef: a tool for testing and resolving phyloreferences
**Submitting author:** @gaurav (<a href="http://orcid.org/0000-0003-0587-0454">Gaurav Vaidya</a>)
**Repository:** <a href="https://github.com/phyloref/jphyloref" target ="_blank">https://github.com/phyloref/jphyloref</a>
**Version:** v1.0.0
