### 2021-06-10 #3206 @kellyrowland
Hi @iimog and @7omasz - would you be available and willing to review this paper for JOSS?
