### 2021-06-11 #3206 @iimog
I'd be willing to review this paper for JOSS. I can start working on the review next week.
