### 2021-06-17 phyloref/jphyloref#82 @iimog
Lots of WARNINGs when building jphyloref
When building `jphyloref` with `mvn package` (and properly setting `JAVA_HOME` #81) there are lots of warnings about "overlapping classes". These are just warnings and building still works but there are so many of them that it is hard to see other build log messages that might be more relevant. In total, 727 of 928 lines in the build log are warnings. Is there a way to suppress (or better avoid) these overlapping class warnings?
