### 2021-06-17 783dbc2
Added some Java installation instructions.
Includes note about JAVA_HOME environment variable as per #81.
