### 2021-06-17 phyloref/jphyloref#81 @gaurav
Hi Markus -- thank you so much for your review and suggestions! I've added a note describing what the `JAVA_HOME` variable needs to be set to in PR #84 -- it's hard to provide instructions for how to do this across all the operating systems, but hopefully this should be enough to clue users in to what might be going wrong and what might fix it. What do you think?
