### 2021-06-17 a3cbaec
Removed or modified tests referring to JFact.
