### 2021-06-17 phyloref/jphyloref#83 @gaurav
Good point! That dependency isn't very useful to us at the moment and doesn't seem to be in active development for OWLAPI 4.x, so it's probably easiest to just remove it entirely (PR #85). This removes the warning you report.
