### 2021-06-17 bbf672c
Recreate JAR file on each run.
This prevents overlapping class warnings from shaded JARs (https://stackoverflow.com/a/11015175/27310).
