### 2021-06-17 phyloref/jphyloref#82 @gaurav
Fixed in bbf672c911a -- turns out this was caused by a couple of inefficient imports and an oddity with the way shaded JARs work. Could you please recheck this with the `improvements-from-joss-review` branch (PR #84)? I'm going to leave that PR open for a bit in case you or our other reviewer have additional issues that need fixing.
