### 2021-06-18 phyloref/jphyloref#81 @iimog
That's perfect, thanks. You definitely can't provide instructions for every platform/OS/version. So having instructions for the main supported ones is fine. I suspect (although I'm biased) that Ubuntu is a rather common choice for users of `jphyloref`. Thanks for adding this note to the README.
