### 2021-06-18 phyloref/jphyloref#82 @iimog
Awesome, can confirm that these WARNINGs are gone in the `improvements-from-joss-review` branch.
