### 2021-06-21 phyloref/jphyloref#87 @iimog
Documentation: statement of need
One of the JOSS review checklist items about documentation is:

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

In my opinion this information is currently lacking from the documentation. At the top the `README`, the [phyloref](https://phyloref.org) website is linked. But for someone not familiar with the topic it is hard to understand what exactly the purpose of the software is. A very brief section at the beginning would be sufficient.
