### 2021-06-21 phyloref/jphyloref#88 @iimog
Documentation: Example usage
One of the JOSS review checklist (https://github.com/openjournals/joss-reviews/issues/3374) items about documentation is:

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There is an example call of `jphyloref` with dummy data in the `README`. However, a couple more examples would be great. This could also be used to demonstrate the kind of problems that are detected by `jphyloref`, i.e. reasons why a phyloreference can not be resolved in the reference phylogeny. The examples need not be real world (in size) but small trees with a hand full of taxa that illustrate what `jphyloref` does help a lot.
