### 2021-06-21 #3374 @iimog
I just finished my initial review:
`jphyloref` is a very useful tool that nicely extends the [family of phyloreferencing tools](https://www.phyloref.org/products/). The software can be easily installed and works as intended. Multiple issues I raised have already been resolved by the authors (https://github.com/phyloref/jphyloref/issues/81, https://github.com/phyloref/jphyloref/issues/82, https://github.com/phyloref/jphyloref/issues/83). Documentation can be slightly improved (https://github.com/phyloref/jphyloref/issues/87, https://github.com/phyloref/jphyloref/issues/88) but beside that I already checked off most of the checklist items.
In the manuscript the "statement of need" section is missing. But most of that information is covered in the "summary" section. Moving this to it's own section with the proper title and adding a bit more "clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience" to the summary would improve comprehensibility.
