### 2021-07-07 175dbf9
Added explanation for purpose of software (as per #87).
