### 2021-07-07 fec0e0f
Removed defunct reasoners; removed redundant link.
