### 2021-07-07 6b93a36
Added reference to the JOSS manuscript.
