### 2021-07-07 741a4ff
Combined command line options into the usage section.
