+++ Usage.md
@@ -1,0 +1,105 @@
+# JPhyloRef Usage
+
+Some example phyloreferences are included as part of the test suite. They can be found in the
+[examples directory], and are tested as part of the [test suite] included with this package.
+You can try these out yourself by setting up [Apache Maven] and running:
+
+```
+$ mvn clean package
+```
+
+This will create `target/JPhyloRef-${version}.jar`. You can run this using Java. For example, the
+following command will provide a comprehensive list of command line options for JPhyloRef.
+
+```
+$ java -jar target/JPhyloRef-1.1.0-SNAPSHOT.jar
+```
+
+You can test one of the example files by running:
+
+```
+$ java -jar target/jphyloref-1.1.0-SNAPSHOT.jar test src/test/resources/phylorefs/dummy1.owl > output.txt
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Input: src/test/resources/phylorefs/dummy1.owl
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Found local ontologies: []
+(...logs and warnings generated by the Elk reasoner are displayed here...)
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Phyloreferences identified: [<#phyloref0>]
+Testing complete:1 successes, 0 failures, 0 failures marked TODO, 0 skipped.
+```
+
+A more detailed description of the test results is written out on the standard output. In this example,
+this is written in the [Test Anything Protocol] (TAP) format to `output.txt`.
+
+```
+1..1
+# From file: src/test/resources/phylorefs/dummy1.owl
+# Using reasoner: null/0.0.0.0
+ok 1 Phyloreference '1'
+# Expected nodes: [#phylogeny0_node2]
+# Resolved nodes: [#phylogeny0_node2]
+```
+
+This indicates that the phyloreference labeled '1' resolved to node 2 on phylogeny 0, which was annotated to
+indicate that this is where this phyloreference was expected resolve. Thus, it was resolved correctly.
+
+Another example file shows how a resolution error is reported.
+
+```
+$ java -jar target/jphyloref-1.1.0-SNAPSHOT.jar test src/test/resources/phylorefs/failing1.jsonld > output.txt
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Input: src/test/resources/phylorefs/failing1.jsonld
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Found local ontologies: []
+(...logs and warnings generated by the Elk reasoner are displayed here...)
+[main] INFO org.phyloref.jphyloref.commands.TestCommand - Phyloreferences identified: [<http://example.org/jphyloref#phyloref0>, <http://example.org/jphyloref#phyloref1>, <http://example.org/jphyloref#phyloref2>]
+Testing complete:1 successes, 1 failures, 0 failures marked TODO, 1 skipped.
+```
+
+Once again, detailed explanations are available on standard output, here redirected to `output.txt`:
+
+```
+1..3
+# From file: src/test/resources/phylorefs/failing1.jsonld
+# Using reasoner: null/0.0.0.0
+ok 1 Phyloreference '1'
+# Expected nodes: [#phylogeny0_node2]
+# Resolved nodes: [#phylogeny0_node2]
+not ok 2 Phyloreference '2'
+# Expected nodes: [#phylogeny0_node1]
+# Resolved nodes: [#phylogeny0_node2]
+# Some nodes were resolved but were not expected: [<http://example.org/jphyloref#phylogeny0_node2>]
+# Some nodes were expected but were not resolved: [<http://example.org/jphyloref#phylogeny0_node1>]
+not ok 3 Phyloreference '4' # SKIP Phyloreference has no expected resolution, and so cannot be tested.
+# Expected nodes: []
[…]
