### 2021-07-07 9de8bce
Attempt to use `console` code blocks; also improved text.
