+++ Usage.md
@@ -38,7 +38,7 @@ ok 1 Phyloreference '1'
 # Resolved nodes: [#phylogeny0_node2]
 ```
 
-This indicates that the phyloreference labeled '1' resolved to node 2 on phylogeny 0, which was annotated to
+This indicates that the phyloreference labeled '1' resolved to node `#phylogeny0_node2`, which was annotated to
 indicate that this is where this phyloreference was expected resolve. Thus, it was resolved correctly.
 
 Another example file shows how a resolution error is reported.
@@ -73,28 +73,29 @@ not ok 3 Phyloreference '4' # SKIP Phyloreference has no expected resolution, an
 
 Here we see that one phyloreference ('1') resolved as expected, another ('2') resolved to a node other than the one it
 was expected to resolve to, and a third ('3') was skipped during testing, as no node had been annotated to indicate that
-it should resolve to it. However, phyloreference '3' did resolve to a node, so -- if that resolution is correct -- then
-it should probably be annotated to indicate that phyloreference '3' should resolve to it.
+it should resolve to it. However, phyloreference '3' did resolve unexpectedly to a node -- this is noted in the output
+as a suggestion that this resolution should be looked into.
 
 Instead of testing the phyloreference, we could also ask JPhyloRef for information on which node each
 phyloreference resolved to:
 
 ```console
-$ java -jar target/jphyloref-1.1.0-SNAPSHOT.jar resolve src/test/resources/phylorefs/failing1.jsonld > output.txt
+$ java -jar target/jphyloref-1.1.0-SNAPSHOT.jar resolve src/test/resources/phylorefs/failing1.jsonld > output.json
 [main] INFO org.phyloref.jphyloref.commands.TestCommand - Input: src/test/resources/phylorefs/failing1.jsonld
 [main] INFO org.phyloref.jphyloref.commands.TestCommand - Found local ontologies: []
 (...logs and warnings generated by the Elk reasoner are displayed here...)
 ```
 
-In this case, a JSON document is written to standard output, showing which node each phyloreference resolved to.
+In this case, a JSON document is written to standard output (here redirected to `output.json`), showing which node each
+phyloreference resolved to.
 
 ```json
 {"phylorefs":{"#phyloref2":["#phylogeny0_node2"],"#phyloref0":["#phylogeny0_node2"],"#phyloref1":["#phylogeny0_node2"]}}
 ```
 
 Real-world example phyloreferences can be found in the [phyx.js repository]. phyx.js includes a
-test suite that includes [jphyloref.js], a Mocha-based test that verifies that these example
-phyloreferences resolve as expected by JPhyloRef.
+test suite that includes [jphyloref.js], a Mocha-based test that uses JPhyloRef to verify that these example
+phyloreferences resolve as expected.
 
 
   [Apache Maven]: https://maven.apache.org/
