### 2021-07-08 65961b5
Update README.md
Added note about other reasoners that might work.

Co-authored-by: Hilmar Lapp <hlapp@drycafe.net>
