+++ paper/paper.md
@@ -26,15 +26,13 @@ date: 15 April 2021
 bibliography: paper.bib
 ---
 
-# Summary
-
-JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of phylogenetic clade definitions, called phyloreferences, and their accompanying reference phylogenetic trees. It has two main goals. The primary one is to facilitate automated testing that the semantics of the logical definitions imply ("resolve to") the correct nodes in the reference tree as clade ancestors. This is key in supporting quality control for the digitization of phylogenetic clade definitions from natural language text to a structured machine-interpretable representation. It also verifies that one of the theoretical foundational premises of phyloreferences, computational reproducibility, holds in practice. The secondary goal is to enable integration with external tools that need to obtain the clade ancestor node(s) resulting from a given ontology of phyloreferences and reference tree(s). When run as part of an automated testing workflow, JPhyloRef reports test results in the cross-platform Test Anything Protocol (TAP) format. When used to find clade ancestor nodes implied by logical clade definitions, results are returned as a JSON object. JPhyloRef uses the OWL API reference library for reading Web Ontology Language (OWL) ontologies, and for the actual ontology reasoning step it uses an external and configurable OWL reasoner.
-
 # Background and Overview
 
 In evolutionary biology, groups of organisms consisting of an ancestor and all of its descendants ("clades") are a fundamental unit for understanding evolution and describing biodiversity [@De_Queiroz2007-xm]. Phylogenetic clade definitions define clades based on shared ancestry, providing the theoretical foundation for the semantics of taxon concepts to be defined and reproducibly resolved within a hypothesis of evolutionary relationships, i.e., a phylogeny [@De_Queiroz1990-ho; @De_Queiroz1992-vy; @De_Queiroz1992-oq]. We have proposed a mechanism, called Phyloreferencing, for representing phylogenetic clade definitions as structured data with fully machine-processable semantics, using the Web Ontology Language (OWL) [@W3C_OWL_Working_Group2012-we]. We refer to such machine-interpretable clade definitions as "phyloreferences" [@Cellinese2021-gp].
 
 Because phyloreferences have an OWL ontology representation in the form of logically defined classes, combining them with an OWL representation of a phylogenetic tree allows an OWL reasoner to infer, based on their logical definitions, which nodes in the phylogeny, if any, instantiate each phyloreference. We refer to this logical inference as "resolving a phyloreference". Phylogenetic clade definitions are normally created and published in the form of natural language text, and in the context of a phylogenetic hypothesis, the so-called reference phylogeny. When digitized, usually through manual curation, to a structured, machine-interpretable representation, which we call a phyloreference, a key question in quality control of both the digitization product and the ontologies and logical expression algorithm it uses arises from one of the founding premises of phyloreferences: given that phyloreferences make the semantics of their represented clade machine-interpretable, can automated computational testing check whether a phyloreference resolves, and only resolves to the node in the reference phylogeny which the original authors of the published phylogenetic clade definition designate as the expected node.
+
+# Commands and design
 
[…]
