+++ paper/paper.md
@@ -32,7 +32,7 @@ In evolutionary biology, groups of organisms consisting of an ancestor and all o
 
 Because phyloreferences have an OWL ontology representation in the form of logically defined classes, combining them with an OWL representation of a phylogenetic tree allows an OWL reasoner to infer, based on their logical definitions, which nodes in the phylogeny, if any, instantiate each phyloreference. We refer to this logical inference as "resolving a phyloreference". Phylogenetic clade definitions are normally created and published in the form of natural language text, and in the context of a phylogenetic hypothesis, the so-called reference phylogeny. When digitized, usually through manual curation, to a structured, machine-interpretable representation, which we call a phyloreference, a key question in quality control of both the digitization product and the ontologies and logical expression algorithm it uses arises from one of the founding premises of phyloreferences: given that phyloreferences make the semantics of their represented clade machine-interpretable, can automated computational testing check whether a phyloreference resolves, and only resolves to the node in the reference phylogeny which the original authors of the published phylogenetic clade definition designate as the expected node.
 
-# Commands and design
+# Commands and Design
 
 To achieve this, we built JPhyloRef, a command-line tool in Java. When run with the `test` command with an OWL ontology containing phyloreferences as well as reference phylogenies as input, JPhyloRef identifies all the phyloreferences within this ontology, compares for each one the inferred phylogeny node(s) to the expected one, and determines success (inferred node identical to expected) or failure (no inferred node, or inferred node different from expected). To better facilitate use in automated continuous integration testing, JPhyloRef also allows for phyloreferences to be marked as draft (resulting in an "incomplete" test status), or as to be skipped (by no nodes having been annotated as an expected resolution of a particular phyloreference). These test results are reported in the Test Anything Protocol (TAP) format (https://testanything.org/), a cross-platform format for reporting test results. JPhyloRef will also return an exit code indicating success (0), failure (the number of failed phyloreferences as a positive integer) or an ontology containing no phyloreferences (-1, interpreted as 255 by POSIX-compatible operating systems).
 
