### 2021-07-17 #3374 @gaurav
Hi @iimog! Thank you so much for the positive review. I think we've addressed all of the concerns you raised, except for your suggestion for rearranging the summary section a bit, which we're currently discussing and will fully address once all the reviews are complete (https://github.com/phyloref/jphyloref/pull/91). Please let us know if there's anything else that needs improving!
