### 2021-07-19 #3374 @kellyrowland
Thanks for disclosing! I think we should err on the side of caution and not have you do the review in that case.
