### 2021-07-19 #3374 @kellyrowland
Great, thank you!

The general reviewer guidelines are at this page: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Review criteria are listed here: https://joss.readthedocs.io/en/latest/review_criteria.html

An example review checklist is here: https://joss.readthedocs.io/en/latest/review_checklist.html

I will have our review bot add you as a reviewer to this paper and I will edit the anchor issue post here to reflect that as well; please accept the invitation to the repository so that you can check off the boxes in the top post in this issue as you progress through your review.

If you have any other questions or run into any issues, feel free to @ me to ask.
