### 2021-07-26 phyloref/jphyloref#92 @jcoliver
Update README execution example
After installing and building with maven, in order to execute, I used:

```
$ java -jar target/jphyloref-1.1.0-SNAPSHOT.jar test src/test/resources/phylorefs/dummy1.owl
```

While the README indicates the JarFile would be `target/jphyloref-1.1.0.jar`. If this is _not_ due to some idiosyncrasy of my system, the README should be updated (note the Usage.md document refers to the JarFile as I saw it on my system, with the SNAPSHOT token).
