### 2021-07-27 phyloref/jphyloref#93 @jcoliver
Consider use of term "clade" vs. "phylogenetic clade"
The term "phylogenetic clade" is a bit of an odd term, as clades are all, by definition, based on phylogenetic relationships.

For example, this sentence (in the README):

> Phylogenetic clade definitions define groups of organisms consisting of an ancestor and all of its descendants ("clades") based on shared ancestry.

Could be written with no loss of information as:

> Clade definitions define groups of organisms consisting of an ancestor and all of its descendants ("clades") based on shared ancestry.

(although dropping the parenthetical "clades" would be warranted in the latter).

Recognizing this term is out there in the literature, I still think the authors could drop the "phylogenetic" from "phylogenetic clade". The alligator in the phyx.js repo implies that JPhyloRef should work just fine with those old, moldy Linnaean names, too.
