### 2021-07-27 #3374 @jcoliver
My initial review echos that of @iimog. This is useful software for phyloreferencing. I also think a clear Statement of Need is warranted. I added a couple of issues: one is minor clarification ([#92](https://github.com/phyloref/jphyloref/issues/92)), while the other might be of a more stylistic bent that does not necessarily need to be immediately addressed ([#93](https://github.com/phyloref/jphyloref/issues/93)).
