### 2021-08-02 ec5c140
Updated README to make version number clearer.
Closes #92.
