### 2021-08-02 0011060
Updated paper to remove unnecesary "phylogenetic".
As suggested in #93.
