+++ paper/paper.md
@@ -28,13 +28,13 @@ bibliography: paper.bib
 
 # Summary
 
-JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of phylogenetic clade definitions, called phyloreferences, and their accompanying reference phylogenetic trees. It has two main goals. The primary one is to facilitate automated testing that the semantics of the logical definitions imply ("resolve to") the correct nodes in the reference tree as clade ancestors. This is key in supporting quality control for the digitization of phylogenetic clade definitions from natural language text to a structured machine-interpretable representation. It also verifies that one of the theoretical foundational premises of phyloreferences, computational reproducibility, holds in practice. The secondary goal is to enable integration with external tools that need to obtain the clade ancestor node(s) resulting from a given ontology of phyloreferences and reference tree(s). When run as part of an automated testing workflow, JPhyloRef reports test results in the cross-platform Test Anything Protocol (TAP) format. When used to find clade ancestor nodes implied by logical clade definitions, results are returned as a JSON object. JPhyloRef uses the OWL API reference library for reading Web Ontology Language (OWL) ontologies, and for the actual ontology reasoning step it uses an external and configurable OWL reasoner.
+JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of clade definitions, called phyloreferences, and their accompanying reference phylogenetic trees. It has two main goals. The primary one is to facilitate automated testing that the semantics of the logical definitions imply ("resolve to") the correct nodes in the reference tree as clade ancestors. This is key in supporting quality control for the digitization of clade definitions from natural language text to a structured machine-interpretable representation. It also verifies that one of the theoretical foundational premises of phyloreferences, computational reproducibility, holds in practice. The secondary goal is to enable integration with external tools that need to obtain the clade ancestor node(s) resulting from a given ontology of phyloreferences and reference tree(s). When run as part of an automated testing workflow, JPhyloRef reports test results in the cross-platform Test Anything Protocol (TAP) format. When used to find clade ancestor nodes implied by logical clade definitions, results are returned as a JSON object. JPhyloRef uses the OWL API reference library for reading Web Ontology Language (OWL) ontologies, and for the actual ontology reasoning step it uses an external and configurable OWL reasoner.
 
 # Background and Overview
 
-In evolutionary biology, groups of organisms consisting of an ancestor and all of its descendants ("clades") are a fundamental unit for understanding evolution and describing biodiversity [@De_Queiroz2007-xm]. Phylogenetic clade definitions define clades based on shared ancestry, providing the theoretical foundation for the semantics of taxon concepts to be defined and reproducibly resolved within a hypothesis of evolutionary relationships, i.e., a phylogeny [@De_Queiroz1990-ho; @De_Queiroz1992-vy; @De_Queiroz1992-oq]. We have proposed a mechanism, called Phyloreferencing, for representing phylogenetic clade definitions as structured data with fully machine-processable semantics, using the Web Ontology Language (OWL) [@W3C_OWL_Working_Group2012-we]. We refer to such machine-interpretable clade definitions as "phyloreferences" [@Cellinese2021-gp].
[…]
