### 2021-08-02 aebe127
Removed unnecessary "phylogenetics" from README.md.
As suggested in #93.
