### 2021-08-07 #3374 @gaurav
@iimog @jcoliver Thank you so much for your reviews! My co-authors and I have made the requested changes in the GitHub issues you opened, and have [updated the manuscript](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03374/joss.03374/10.21105.joss.03374.pdf) as per your suggestions. Please let us know if you have any further suggestions for improvement!
