### 2021-08-08 19f329f
Updated CHANGELOG for v1.1.0.
