+++ paper/paper.bib
@@ -25,10 +25,10 @@
               definitions transformed into formal logic expressions. We call
               such expressions phyloreferences, and argue that, unlike Linnaean
               names, they meet all requirements for effective data-integration.",
-  journal  = "EcoEvoRxiv",
-  month    =  mar,
-  year     =  2021,
-  doi      =  {10.32942/osf.io/57yjs}
+  journal  = "Philosophy, Theory and Practice in Biology",
+  year     = {2021},
+  pages    = "Accepted",
+  url      = {https://doi.org/10.32942/osf.io/57yjs}
 }
 
 @ARTICLE{De_Queiroz1992-oq,
