+++ paper/paper.bib
@@ -27,7 +27,7 @@
               names, they meet all requirements for effective data-integration.",
   journal  = "Philosophy, Theory and Practice in Biology",
   year     = {2021},
-  pages    = "Accepted",
+  pages    = {Accepted. Preprint available on {EcoEvoXiv}},
   url      = {https://doi.org/10.32942/osf.io/57yjs}
 }
 
