+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Cellinese2021-gp,
-  title    = "Phyloreferences: {{Tree-Native}}, Reproducible, and
-              {Machine-Interpretable} Taxon Concepts",
+  title    = "Phyloreferences: {Tree-Native}, {Reproducible}, and
+              {Machine-Interpretable} {Taxon} {Concepts}",
   author   = "Cellinese, Nico and Conix, Stijn and Lapp, Hilmar",
   abstract = "Evolutionary and organismal biology, similar to other fields in
               biology, have become inundated with data. At the same rate, we
