+++ paper/paper.bib
@@ -27,7 +27,7 @@
               names, they meet all requirements for effective data-integration.",
   journal  = "Philosophy, Theory and Practice in Biology",
   year     = {2021},
-  pages    = {Accepted. Preprint available on {EcoEvoXiv}},
+  pages    = {Accepted. Preprint available on {EcoEvoRxiv}},
   url      = {https://doi.org/10.32942/osf.io/57yjs}
 }
 
