+++ paper/paper.bib
@@ -27,8 +27,7 @@
               names, they meet all requirements for effective data-integration.",
   journal  = "Philosophy, Theory and Practice in Biology",
   year     = {2021},
-  pages    = {Accepted. Preprint available on {EcoEvoRxiv}},
-  url      = {https://doi.org/10.32942/osf.io/57yjs}
+  pages    = {Accepted. Preprint available on {EcoEvoRxiv} at \url{https://doi.org/10.32942/osf.io/57yjs}},
 }
 
 @ARTICLE{De_Queiroz1992-oq,
