### 2021-08-09 phyloref/jphyloref#99 @jcoliver
Minor fixes to paper/paper.md
+ First sentence of summary, make "clade" plural and drop second instance of "definitions":
    + "JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of clade**s** ~definitions~, called phyloreferences"
+ Line 44, add question mark at end of question:
    + given that phyloreferences render the semantics of their represented clades machine-interpretable, can a computational algorithm test whether a phyloreference resolves, and only resolves to the node in the reference phylogeny that the original authors of the published clade definition designate as the expected node **?**
