### 2021-08-09 #3374 @jcoliver
Great! I suggested [two minor fixes](https://github.com/phyloref/jphyloref/issues/99) to the manuscript article, but everything else on the checklist looks good to me.
