### 2021-08-09 65fb22e
Incremented development version number to `1.2.0`.
