### 2021-08-10 #3374 @kellyrowland
Thank you @jcoliver !

@gaurav please let me know about the minor fixes, and once you have gotten through them, please:

- issue a new tagged release of the software (if it has changed in the review process),
- archive the release on Zenodo, figshare, or another site, and
- post the version number and archive DOI here in this review issue

Feel free to ping me with any questions!
