### 2021-08-10 98500dd
Fixes minor errors resulting from previous edits
Fixes #99.
