+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of clade definitions, called phyloreferences, and their accompanying reference phylogenetic trees. It has two main goals. The primary one is to facilitate automated testing that the semantics of the logical definitions imply ("resolve to") the correct nodes in the reference tree as clade ancestors. The secondary goal is to enable integration with external tools that need to obtain the clade ancestor node(s) resulting from a given ontology of phyloreferences and reference tree(s). When run as part of an automated testing workflow, JPhyloRef reports test results in the cross-platform Test Anything Protocol (TAP) format. When used to find clade ancestor nodes implied by logical clade definitions, results are returned as a JSON object. JPhyloRef uses the OWL API reference library for reading Web Ontology Language (OWL) ontologies, and for the actual ontology reasoning step it uses an external and configurable OWL reasoner.
+JPhyloRef is a command line tool as well as a web service for reasoning with ontologies containing logical definitions of clades, called phyloreferences, and their accompanying reference phylogenetic trees. It has two main goals. The primary one is to facilitate automated testing that the semantics of the logical definitions imply ("resolve to") the correct nodes in the reference tree as clade ancestors. The secondary goal is to enable integration with external tools that need to obtain the clade ancestor node(s) resulting from a given ontology of phyloreferences and reference tree(s). When run as part of an automated testing workflow, JPhyloRef reports test results in the cross-platform Test Anything Protocol (TAP) format. When used to find clade ancestor nodes implied by logical clade definitions, results are returned as a JSON object. JPhyloRef uses the OWL API reference library for reading Web Ontology Language (OWL) ontologies, and for the actual ontology reasoning step it uses an external and configurable OWL reasoner.
 
 # Background
 
@@ -38,7 +38,7 @@ Because phyloreferences have an OWL ontology representation in the form of logic
 
 # Statement of need
 
-Clade definitions are normally created and published in the form of natural language text, and in the context of a phylogenetic hypothesis, represented by the so-called reference phylogeny. When digitized to a phyloreference, usually through a process of manual curation, a key question in quality control of the digitization product as well as the involved ontologies and the generated logical expressions arises from one of the founding premises of phyloreferences: given that phyloreferences render the semantics of their represented clades machine-interpretable, can a computational algorithm test whether a phyloreference resolves, and only resolves to the node in the reference phylogeny that the original authors of the published clade definition designate as the expected node. We built JPhyloRef to enable this type of test, and to allow integrating such tests into automated testing frameworks. We also needed other tools to programmatically answer for a given ontology of phyloreferences and an accompanying phylogeny, which phyloreferences resolve to which nodes (as clade ancestors) in the phylogeny.
+Clade definitions are normally created and published in the form of natural language text, and in the context of a phylogenetic hypothesis, represented by the so-called reference phylogeny. When digitized to a phyloreference, usually through a process of manual curation, a key question in quality control of the digitization product as well as the involved ontologies and the generated logical expressions arises from one of the founding premises of phyloreferences: given that phyloreferences render the semantics of their represented clades machine-interpretable, can a computational algorithm test whether a phyloreference resolves, and only resolves to the node in the reference phylogeny that the original authors of the published clade definition designate as the expected node? We built JPhyloRef to enable this type of test, and to allow integrating such tests into automated testing frameworks. We also needed other tools to programmatically answer for a given ontology of phyloreferences and an accompanying phylogeny, which phyloreferences resolve to which nodes (as clade ancestors) in the phylogeny.
 
 # JPhyloRef feature overview
 
