+++ paper/paper.bib
@@ -186,11 +186,11 @@
 @software{gaurav_vaidya_2021_4697965,
   author       = {Gaurav Vaidya and
                   Hilmar Lapp},
-  title        = {phyloref/jphyloref: JPhyloRef v1.0.0},
-  month        = apr,
+  title        = {{phyloref}/{jphyloref}: JPhyloRef v1.1.0},
+  month        = aug,
   year         = 2021,
   publisher    = {Zenodo},
-  version      = {v1.0.0},
-  doi          = {10.5281/zenodo.4697965},
-  url          = {https://doi.org/10.5281/zenodo.4697965}
+  version      = {v1.1.0},
+  doi          = {10.5281/zenodo.5172998},
+  url          = {https://doi.org/10.5281/zenodo.5172998}
 }
