### 2021-08-10 7117c98
Updated CHANGELOG to include v1.1.1.
