+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Based on the suggestion at https://keepachangelog.com/en/1.0.0/.
 
 ## [Unreleased]
 
+## [1.1.1] - 2021-08-10
+- Fixed manuscript for submission to the Journal of Open Source Software (JOSS).
+
 ## [1.1.0] - 2021-08-07
 - Prepared manuscript for submission to the Journal of Open Source Software (JOSS).
 - Improved documentation based on JOSS reviewer feedback.
@@ -63,7 +66,8 @@ Based on the suggestion at https://keepachangelog.com/en/1.0.0/.
 - Initial release, with support for testing phyloreferences expressed in OWL
   and stored in RDF/XML.
 
-[Unreleased]: https://github.com/phyloref/jphyloref/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/phyloref/jphyloref/compare/v1.1.1...HEAD
+[1.1.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.1.1
 [1.1.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.1.0
 [1.0.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.0.0
 [0.4.0]: https://github.com/phyloref/jphyloref/releases/tag/v0.4.0
