### 2021-08-10 130e60c
Fixed typo in CHANGELOG.
