+++ CHANGELOG.md
@@ -67,7 +67,7 @@ Based on the suggestion at https://keepachangelog.com/en/1.0.0/.
   and stored in RDF/XML.
 
 [Unreleased]: https://github.com/phyloref/jphyloref/compare/v1.1.1...HEAD
-[1.1.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.1.1
+[1.1.1]: https://github.com/phyloref/jphyloref/releases/tag/v1.1.1
 [1.1.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.1.0
 [1.0.0]: https://github.com/phyloref/jphyloref/releases/tag/v1.0.0
 [0.4.0]: https://github.com/phyloref/jphyloref/releases/tag/v0.4.0
