### 2021-08-11 #3374 @gaurav
@kellyrowland All done!
* The latest released version of JPhyloRef is [v1.1.1](https://github.com/phyloref/jphyloref/releases/tag/v1.1.1).
* This has been archived to Zenodo at [doi:10.5281/zenodo.5178723](https://doi.org/10.5281/zenodo.5178723).
* I've updated the manuscript so that it cites the archived version of the software in Zenodo.

Thank you everybody!
